+++ src/pages/favs.tsx
@@ -1,0 +1,5 @@
+import Favs from '../components/favs';
+
+export const FavsPage: React.FC = () => {
+  return <Favs />;
+};